backend/requirements.txt
requests==2.26.0
django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser
# ? django-filter==2.4.0
Pillow==8.3.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv
# ? pytz
# ? sqlparse==0.4.2